fix: rename functions related to funnelStep
