+++ src/constants/funnelStep.ts
@@ -1,0 +1,3 @@
+export const linkStepNames = ['입력전', '링크입력', '추출상태'] as const;
+export const extractedPlaceStepNames = ['추출장소', '리스트'] as const;
+export const searchPlaceStepNames = ['추출장소', '리스트선택', '리스트'] as const;
