--- src/constants/linkForm.ts
@@ -1,1 +1,0 @@
-export const linkStepNames = ['입력전', '링크입력', '추출상태'] as const;
